# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    CodeCloud API. Primary authentication uses Clerk (session cookies managed
    by @clerk/express middleware). The /auth/* JWT endpoints are an optional
    secondary auth layer for programmatic API clients that cannot use Clerk
    sessions.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/collaborators:
    get:
      operationId: listCollaborators
      tags: [collaborators]
      summary: List project collaborators
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collaborators list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collaborator"
    post:
      operationId: addCollaborator
      tags: [collaborators]
      summary: Add collaborator to project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCollaboratorBody"
      responses:
        "201":
          description: Added collaborator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collaborator"
  /projects/{projectId}/collaborators/{userId}:
    delete:
      operationId: removeCollaborator
      tags: [collaborators]
      summary: Remove collaborator
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Collaborator removed
...
components:
  schemas:
...
    Collaborator:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        userId:
          type: string
        role:
          type: string
          enum: [viewer, editor, admin]
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - userId
        - role
        - createdAt
    AddCollaboratorBody:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [viewer, editor, admin]
          default: editor
      required:
        - username
...